requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
albumentations>=1.3.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
tqdm>=4.65.0
pillow>=9.0.0
pandas>=1.5.0
pyyaml>=6.0
datasets>=2.14.0
huggingface-hub>=0.16.0 